% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_where}
\alias{cat_where}
\title{cat function for whereami}
\usage{
cat_where(
  where,
  type = c("rule", "boxx", "bullet", "line", "print"),
  color = c("none", "black", "blue", "cyan", "green", "magenta", "red", "white",
    "yellow", "grey", "silver"),
  ...
)
}
\arguments{
\item{where}{whereami output}

\item{type}{\code{cli} \verb{cat_*} method. Can be \code{boxx}, \code{bullet}, \code{line}, \code{print}, \code{rule}.
Default is \code{rule}.}

\item{color}{\code{cli} \verb{col_*} method.}

\item{...}{further arguments passed to \verb{cli::cat_*}.}
}
\value{
character
}
\description{
cat function for whereami output
}
\examples{
 cat_where(whereami())
 cat_where(whereami(), type = "bullet", bullet = "tick")
 cat_where(whereami(), color = "green")
}
\author{
Colin Fay
}
\concept{print}
