% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter_state}
\alias{counter_state}
\alias{counter_reset}
\title{whereami counter}
\usage{
counter_state()

counter_reset()
}
\description{
Interact with internal whereami counter
}
\details{
counter_state returns current hit count for each stored counter,
arranged in order of counter initialization.

counter_reset will remove all stored counters.
}
