% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{set_whereami_log}
\alias{set_whereami_log}
\title{Set Log Path}
\usage{
set_whereami_log(path = tempdir())
}
\arguments{
\item{path}{character, path to write to, Default: tempdir()}
}
\description{
Set the path to write the counter log as a json
}
