% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereami.R
\name{whereami}
\alias{whereami}
\title{Return where the script is run from}
\usage{
whereami(path_expand = FALSE, counter = FALSE)
}
\arguments{
\item{path_expand}{logical, expand relational path, Default: FALSE}

\item{counter}{logical, remember where whereami last ran and increase the
internal counter by 1, Default: FALSE}
}
\value{
character
}
\description{
Returns the file that a command is run from. If traceback is available
  then the line that it was run from is also returned.
}
\details{
does not work currently on rmd render
}
\examples{
\dontrun{
if(interactive()){
 whereami()
 whereami(path_expand = TRUE)
 }
}
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
