% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{cat_where}
\alias{cat_where}
\title{cat function for whereami}
\usage{
cat_where(where)
}
\arguments{
\item{where}{whereami output}
}
\value{
character
}
\description{
cat function for whereami output
}
\examples{
\dontrun{
 cat_where(whereami())
}
}
\concept{print}
