% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter_state}
\alias{counter_state}
\alias{counter_reset}
\alias{counter_names}
\alias{counter_get}
\alias{autoplot.whereamicounter}
\title{whereami counter}
\usage{
counter_state(item = NULL)

counter_reset(item = NULL)

counter_names()

counter_get()

autoplot.whereamicounter(counter)
}
\arguments{
\item{item}{character, name of the counter to access, Default: NULL}

\item{counter}{For autoplot, the output of \code{counter_get()}}
}
\description{
Interact with internal whereami counter
}
\details{
\code{counter_names()} returns names of the active stored counters.

\code{counter_state()} returns current hit count for item, if NULL then all counters
are returned.

\code{counter_reset()} will remove counter of item, if item is NULL then all counters
are reset.

\code{counter_get()} returns a table with all the counters. These can then
be plotted with the \code{autoplot()} method.
}
\concept{counter}
