% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereami.R
\name{whereami}
\alias{whereami}
\title{Return where the script is run from}
\usage{
whereami(path_expand = FALSE, tag = NULL)
}
\arguments{
\item{path_expand}{logical, expand relational path, Default: FALSE}

\item{tag}{character, optional tag for the call, Default: NULL}
}
\value{
character
}
\description{
Returns the file that a command is run from.

If traceback is available then the line that it was run from is also returned.
}
\examples{
 whereami()
 whereami(path_expand = TRUE)
}
\author{
Jonathan Sidi
}
\concept{main}
