% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereami.R
\name{whereami}
\alias{whereami}
\title{FUNCTION_TITLE}
\usage{
whereami(path_expand = FALSE)
}
\arguments{
\item{path_expand}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rstudioapi]{rstudio-editors}}
}
